function-app/fetch_data/requirements.txt
azure-functions
requests
azure-cosmos
azure-ai-vision-imageanalysis
azure-ai-documentintelligence
python-dotenv

